refactor(types): radio-group
